add admin order history apis

+++ jcart-administration-back/src/main/java/com/yh/jcartadministrationback/po/OrderHistory.java
@@ -2,11 +2,10 @@ package com.yh.jcartadministrationback.po;
 
 import java.util.Date;
 
-//订单历史记录
 public class OrderHistory {
     private Long orderHistoryId;
 
-    private Integer orderId;
+    private Long orderId;
 
     private Date time;
 
@@ -24,11 +23,11 @@ public class OrderHistory {
         this.orderHistoryId = orderHistoryId;
     }
 
-    public Integer getOrderId() {
+    public Long getOrderId() {
         return orderId;
     }
 
-    public void setOrderId(Integer orderId) {
+    public void setOrderId(Long orderId) {
         this.orderId = orderId;
     }
